layer2/face_server/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.9
numpy<2.0.0
opencv-python-headless==4.8.1.78
Pillow==10.0.0
python-jose==3.3.0
python-dotenv==1.0.0
# dlib==19.24.99
tensorflow==2.14.0
facenet-pytorch==2.5.3
slowapi==0.1.8